requirements.txt
# Core Framework
Flask==3.0.0
python-dotenv==1.0.0
Werkzeug==3.0.1

# Database
Flask-SQLAlchemy==3.1.1
Flask-Migrate==4.0.5
psycopg2-binary==2.9.9

# Authentication & Security
Flask-Login==0.6.3
Flask-JWT-Extended==4.6.0
Flask-Bcrypt==1.0.1
Flask-WTF==1.2.1
WTForms==3.1.1
Flask-Limiter==3.5.0
Flask-CORS==4.0.0
email-validator==2.1.0

# Real-time
Flask-SocketIO==5.3.6
python-socketio==5.11.0
python-engineio==4.9.0

# Email
Flask-Mail==0.9.1

# Caching & Session
redis==5.0.1
Flask-Session==0.6.0

# ML & Data Science
numpy==1.26.3
pandas==2.1.4
scikit-learn==1.3.2
joblib==1.3.2
xgboost==2.0.3
shap==0.44.0
imbalanced-learn==0.11.0

# Visualization
matplotlib==3.8.2
seaborn==0.13.1

# Export & Reports
reportlab==4.0.8
openpyxl==3.1.2
fpdf2==2.7.7
Pillow==10.3.0

# API Documentation
flask-swagger-ui==4.11.1
apispec==6.3.1
apispec-webframeworks==1.0.0

# Background Tasks
celery==5.3.4

# Monitoring
psutil==5.9.7

# HTTP Requests (for webhooks)
requests==2.31.0

# Utilities
python-dateutil==2.8.2
pytz==2024.1

# Testing
pytest==7.4.3
pytest-cov==4.1.0
pytest-flask==1.3.0
faker==22.5.1

# Production
gunicorn==22.0.0

# Code Quality
flake8==7.0.0
black==24.1.1
